removes formatting from file based navigation ad renames map-component to acfurately reflect role of file
